flash/requirements.txt
Flask
Jinja2
pillow
requests
